feat: add typ in feature
