+++ src/feature/comment/type.ts
@@ -1,0 +1,5 @@
+export interface NewComment {
+    body: string
+    postId: number | null
+    userId: number
+  }
