+++ src/feature/table/type.ts
@@ -1,0 +1,7 @@
+import { Post as originPost } from "../../entities"
+import { Author } from "../../shared/types"
+
+
+export interface Post extends originPost {
+  author: Author
+}
