+++ src/shared/types/index.ts
@@ -1,10 +1,3 @@
-import { Post as originPost } from "../../entities"
-
-
-export interface Post extends originPost {
-  author: Author
-}
-
 export interface Author {
   id: number
   username: string
@@ -15,10 +8,4 @@ export interface Pagination {
   limit: number
   skip: number
   total: number
-}
-
-export interface NewComment {
-  body: string
-  postId: number | null
-  userId: number
 }
